requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
langchain-google-genai>=1.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
